DRY up common code between BSD / OSX.
